# api/adapters/http/contract/models.py
"""REST API data models."""

# generated by fastapi-codegen:
#   filename:  openapi.yml
#   timestamp: 2025-04-06T12:21:27+00:00

from uuid import UUID

from pydantic import BaseModel, Field, RootModel


class Location(BaseModel):
    x: int = Field(..., description="X")
    y: int = Field(..., description="Y")


class Order(BaseModel):
    id: UUID = Field(..., description="Идентификатор")
    location: Location = Field(..., description="Геолокация")


class Courier(BaseModel):
    id: UUID = Field(..., description="Идентификатор")
    name: str = Field(..., description="Имя")
    location: Location = Field(..., description="Геолокация")


class Error(BaseModel):
    code: int = Field(..., description="Код ошибки")
    message: str = Field(..., description="Текст ошибки")


class ApiV1OrdersActiveGetResponse(RootModel[list[Order]]):
    root: list[Order]


class ApiV1CouriersGetResponse(RootModel[list[Courier]]):
    root: list[Courier]